NeuroTwin-Final/requirements.txt
streamlit
pandas
numpy
plotly
pyttsx3
fpdf2
kaleido
streamlit-components
